second pass at scheduler, need to update model

+++ sched/app_api.h
@@ -5,6 +5,6 @@
 #include "core/types.h"
 
 u8 sched_app_publish_OUT_event(u8* data, u32 size, u32 driverID);
-u8 sched_app_pop_IN_event(u32 driverID, u8* buffer, u32 size);
+u32 sched_app_pop_IN_event(u32 driverID, u8* buffer, u32 size);
 
 #endif
